chore: remove old help command
